nightmare: deal with dying upstreams (before response)
Upstream servers may die unexpectedly due to apps lacking
proper timeout mechanisms.  We'll try to return a 502 error
in that case.

There isn't much we can do if they die in the middle of writing
a response, but the chance of that is smaller than dying before
any response is written.  Parsing the entire response body
would be too much for a rare case (and keepalive_timeout will
expire clients, anyways).